Renamed Button2.py to button2.py
Changed the Button2.py name to button2.py.

Added in the buttons which are loaded in by a for loop, this for loop loads images on a range of the pic_list that I stated before.
I then played around with the numbers  in tile_button so they display in a sorted manner.

I draw the buttons and added a rectangle to cover the extra lines for a better looking ui.