Added Mouse control/positions
Added a way to get the mouse position.

Added a if statement to check if the mouse is inside the area where we place the tiles.

Mapped the left click to update the world_data tile when clicked (Add tile)

Mapped mouse right click to reset world_data selected address to -1 when pressed (remove tile)