Level change by arrows
Added another draw_txt line which explains that to change lvl you need to press up or down.

Also added keyboard input if the key up is pressed to =+1
And if the key down is pressed to -+ 1 if the value is higher than 0. (to not get into negative numbers).

I also changed the font size to fit the text better in the screen and filled the background in PURPLE.