Made it so 1 lvl made in editor loads in the main game.
After some time working in a separate fork.
I managed to realize I can load lvls in lvl_data.py.
Imported pickle in lvl_data.py, used pickle_in to read the data in Cyber_level1_data as binary and store it in world_data.

Resized the main game file to be more friendly to editor levels.

Changed ground value so 1 are pun instead of 0s on the worldmap.